+++ CHANGELOG.md
@@ -7,7 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+- [Added dynamic gas station metadata fetch](https://github.com/multiversx/mx-sdk-dapp/pull/1398)
+
 ## [[v4.0.1](https://github.com/multiversx/mx-sdk-dapp/pull/1397)] - 2025-03-31
+
 - [Added shard data to the Ledger address table](https://github.com/multiversx/mx-sdk-dapp/pull/1396)
 
 ## [[v4.0.0](https://github.com/multiversx/mx-sdk-dapp/pull/1395)] - 2025-03-27